bot/requirements.txt
python-telegram-bot[async]
asyncpg
qrcode
fastapi
uvicorn